requirements.txt
google-cloud-aiplatform==1.88.0
python-dotenv==1.1.0
anthropic==0.6.1
pandas==2.2.3
sqlglot==26.13.0
func_timeout==4.3.5
sqlvalidator==0.0.20
tqdm==4.67.1
datasketch==1.6.5
langchain-google-vertexai==2.0.20